fix(ui): Keep the multi-select list clear of the mobile keyboard
Changes:
- Render the list inline below md instead of in a popover, so it sits in normal flow
- Cap the popover to Radix's available height and add collision padding above md
- Split multi-select.tsx into prefixed sibling files behind the same import path
- Filter and rank options case- and diacritic-blind, reusing the products tier ladder
- Extract that ladder into utils/search-relevance.ts so products and pickers share one algorithm
- Pass chain labels as cmdk keywords, since those options are valued by code
- Make the modal body the only flexible track, so large text scrolls instead of squashing the header and footer
- Give .CollapsibleContent a prefers-reduced-motion opt-out

The list was anchored to a trigger inside a container that shrinks to a couple of hundred pixels when the keyboard opens, while asking for a flat 300px, so collision handling had no room to work with. In flow there is no anchor and no portal, and the browser scrolls the focused field into view by itself. Search separately ignored diacritics, so "baska" never offered "Baska" with the caron.

Notes:
Escape needs a window-level capture listener: Radix's dismissable layers claim Escape in the capture phase on document and only the topmost layer acts, and an inline list is not one of those layers, so a dialog around it would otherwise close instead.
