+++ frontend/src/app/products/utils/product-relevance.ts
@@ -1,16 +1,6 @@
 import type { ProductResponse } from "@/lib/cijene-api/schemas";
 import { normalizeForSearch } from "@/utils/strings";
-
-/** How well the whole query sits in the text, best tier first. */
-const MATCH_TIER = {
-  exact: 0,
-  prefix: 1,
-  wordPrefix: 2,
-  contains: 3,
-  none: 4,
-};
-
-const NO_POSITION = Number.MAX_SAFE_INTEGER;
+import { findWholeQuery } from "@/utils/search-relevance";
 
 interface IProductRelevance {
   product: ProductResponse;
@@ -19,20 +9,6 @@ interface IProductRelevance {
   position: number;
   chainCount: number;
   nameLength: number;
-}
-
-function findWholeQuery(haystack: string, query: string) {
-  if (haystack === query) return { tier: MATCH_TIER.exact, position: 0 };
-
-  const position = haystack.indexOf(query);
-  if (position === -1) return { tier: MATCH_TIER.none, position: NO_POSITION };
-  if (position === 0) return { tier: MATCH_TIER.prefix, position };
-
-  const startsWord = /[^a-z0-9]/.test(haystack[position - 1]);
-  return {
-    tier: startsWord ? MATCH_TIER.wordPrefix : MATCH_TIER.contains,
-    position,
-  };
 }
 
 function findBestWholeQuery(name: string, brand: string, query: string) {
